:recycle: move `PrepareDecorator`
