+++ src/server_prepare/configure.rs
@@ -1,6 +1,5 @@
 use std::error;
 use std::fmt::Display;
-use std::future::Future;
 use std::net::SocketAddr;
 
 use hyper::server::accept::Accept;
@@ -8,7 +7,6 @@ use hyper::server::conn::AddrIncoming;
 use hyper::server::Builder;
 use hyper::Server;
 
-use crate::PrepareError;
 
 /// binding to any kind of income stream
 ///
@@ -78,33 +76,4 @@ where
     }
 }
 
-/// add decorator for each prepare 's [`Future`]
-///
-///It is useful for adding extra functional on original prepare task
-pub trait PrepareDecorator: 'static {
-    type OutFut<'fut, Fut, T>: Future<Output = Result<T, PrepareError>> + 'fut
-    where
-        Fut: Future<Output = Result<T, PrepareError>> + 'fut,
-        T: 'static;
-
-    fn decorator<'fut, Fut, T>(in_fut: Fut) -> Self::OutFut<'fut, Fut, T>
-    where
-        Fut: Future<Output = Result<T, PrepareError>> + 'fut,
-        T: 'static;
-}
-
-/// Default Decorator without any effect
-pub struct EmptyDecorator;
-
-impl PrepareDecorator for EmptyDecorator {
-    type OutFut<'fut, Fut, T> = Fut
-        where Fut: Future<Output=Result<T, PrepareError>> + 'fut  ,T: 'static;
-
-    fn decorator<'fut, Fut, T>(in_fut: Fut) -> Self::OutFut<'fut, Fut, T>
-    where
-        Fut: Future<Output = Result<T, PrepareError>> + 'fut,
-        T: 'static,
-    {
-        in_fut
-    }
-}
+pub use super::decorator::{PrepareDecorator,EmptyDecorator};
