+++ src/server_prepare/decorator.rs
@@ -1,5 +1,5 @@
-use crate::server_prepare::PrepareDecorator;
-use crate::ServerPrepare;
+use std::future::Future;
+use crate::{PrepareError, ServerPrepare};
 
 impl<C, Effect, Log, State, Graceful, Decorator>
     ServerPrepare<C, Effect, Log, State, Graceful, Decorator>
@@ -13,5 +13,36 @@ impl<C, Effect, Log, State, Graceful, Decorator>
         D: PrepareDecorator,
     {
         ServerPrepare::new(self.prepares.change_decorator(), self.graceful, self.span)
+    }
+}
+
+/// add decorator for each prepare 's [`Future`]
+///
+///It is useful for adding extra functional on original prepare task
+pub trait PrepareDecorator: 'static {
+    type OutFut<'fut, Fut, T>: Future<Output = Result<T, PrepareError>> + 'fut
+        where
+            Fut: Future<Output = Result<T, PrepareError>> + 'fut,
+            T: 'static;
+
+    fn decorator<'fut, Fut, T>(in_fut: Fut) -> Self::OutFut<'fut, Fut, T>
+        where
+            Fut: Future<Output = Result<T, PrepareError>> + 'fut,
+            T: 'static;
+}
+
+/// Default Decorator without any effect
+pub struct EmptyDecorator;
+
+impl PrepareDecorator for EmptyDecorator {
+    type OutFut<'fut, Fut, T> = Fut
+        where Fut: Future<Output=Result<T, PrepareError>> + 'fut  ,T: 'static;
+
+    fn decorator<'fut, Fut, T>(in_fut: Fut) -> Self::OutFut<'fut, Fut, T>
+        where
+            Fut: Future<Output = Result<T, PrepareError>> + 'fut,
+            T: 'static,
+    {
+        in_fut
     }
 }
